# openapi.json
{
  "components": {
    "schemas": {
      "Body_create_download_form_web_downloads_post": {
        "properties": { "url": { "maxLength": 2000, "title": "Url", "type": "string" } },
        "required": ["url"],
        "title": "Body_create_download_form_web_downloads_post",
        "type": "object"
      },
...
      "Body_login_form_web_login_post": {
        "properties": {
          "email": { "maxLength": 255, "title": "Email", "type": "string" },
          "password": { "maxLength": 255, "title": "Password", "type": "string" },
          "return_url": {
            "anyOf": [{ "maxLength": 500, "type": "string" }, { "type": "null" }],
            "title": "Return Url"
          }
        },
        "required": ["email", "password"],
        "title": "Body_login_form_web_login_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": { "title": "Context", "type": "object" },
          "input": { "title": "Input" },
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "title": "Location",
            "type": "array"
          },
          "msg": { "title": "Message", "type": "string" },
          "type": { "title": "Error Type", "type": "string" }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": { "HTTPBearer": { "scheme": "bearer", "type": "http" } }
  },
  "info": {
    "contact": { "name": "Team 21", "url": "https://github.com/tomkabel/team21-vooglaadija" },
    "description": "REST API for user authentication, creating download jobs, tracking job status, and retrieving processed files. Authentication uses bearer JWT access tokens.",
    "license": { "name": "GPLv3", "url": "https://www.gnu.org/licenses/gpl-3.0.html" },
    "summary": "Asynchronous API for authenticated video download jobs.",
    "title": "Vooglaadija API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/web/downloads": {
      "get": {
        "description": "Render main dashboard page with download list.",
        "operationId": "dashboard_page_web_downloads_get",
        "responses": {
          "200": {
            "content": { "application/json": { "schema": {} } },
            "description": "Successful Response"
          }
        },
        "security": [{ "HTTPBearer": [] }],
        "summary": "Dashboard Page",
        "tags": ["web"]
      },
      "post": {
        "description": "HTMX endpoint for form submissions. Returns HTML fragment.",
        "operationId": "create_download_form_web_downloads_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_download_form_web_downloads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": { "application/json": { "schema": {} } },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [{ "HTTPBearer": [] }],
        "summary": "Create Download Form",
        "tags": ["web"]
      }
    },
...
    "/web/downloads/{job_id}": {
      "delete": {
        "description": "HTMX endpoint for deleting a download.",
        "operationId": "delete_download_form_web_downloads__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": { "title": "Job Id", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": { "application/json": { "schema": {} } },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [{ "HTTPBearer": [] }],
        "summary": "Delete Download Form",
        "tags": ["web"]
      }
    },
...
    "/web/login": {
      "get": {
        "description": "Render login page.",
        "operationId": "login_page_web_login_get",
        "parameters": [
          {
            "in": "query",
            "name": "return_url",
            "required": false,
            "schema": { "default": "/web/downloads", "title": "Return Url", "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": { "application/json": { "schema": {} } },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login Page",
        "tags": ["web"]
      },
      "post": {
        "description": "Handle login form submission via HTMX or regular POST.",
        "operationId": "login_form_web_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": { "$ref": "#/components/schemas/Body_login_form_web_login_post" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": { "application/json": { "schema": {} } },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login Form",
        "tags": ["web"]
      }
    },
...
  },
...
}